port theme to common/ui - adjust references to theme in GO and SRO frontends, and dedupe as necessary - rename brand[GO|SRO]500 to brand500 - remove extra theme file in SRO frontend
