+++ apps/gi-frontend/src/app/components/Header.tsx
@@ -106,7 +106,7 @@ function HeaderContent() {
                 backgroundColor: 'rgba(255,255,255,0.1)',
               },
               '& .Mui-selected': {
-                backgroundImage: `linear-gradient(to top, ${theme.palette.brandGO500.main}, ${theme.palette.neutral700.main})`,
+                backgroundImage: `linear-gradient(to top, ${theme.palette.brand500.main}, ${theme.palette.neutral700.main})`,
                 color: `${theme.palette.neutral100.main} !important`,
                 textShadow:
                   '0.25px 0 0 currentColor, -0.25px 0 0 currentColor, 0 0.25px 0 currentColor, 0 -0.25px 0',
@@ -125,7 +125,7 @@ function HeaderContent() {
                     sx={(theme) => ({
                       height: '24px',
                       width: '24px',
-                      boxShadow: `0 0 10px 1px ${theme.palette.brandGO500.main}`,
+                      boxShadow: `0 0 10px 1px ${theme.palette.brand500.main}`,
                     })}
                   />
                   <Typography variant="h6" sx={{ px: 1, fontWeight: 'Normal' }}>
